admin-microservice-dev/requirements.txt
aiohttp
aiosignal
aniso8601
anyio
apiclient4
argon2-cffi
argon2-cffi-bindings
async-generator
async-timeout
attrs
cachetools
certifi
cffi
charset-normalizer
click
colorama
cryptography
cssselect
dateparser
distlib
docx
et-xmlfile
Faker
fastapi
feedfinder2
feedparser
filelock
flasgger
Flask
Flask-Cors
Flask-RESTful
flatten-json
frozenlist
h11
httpcore
httplib2
httpx
idna
importlib-metadata
itsdangerous
jieba3k
Jinja2
joblib
jsonschema
jwt
lockfile
lxml
MarkupSafe
mistune
msgpack
multidict
mysql-connector-python
netaddr
newspaper3k
nltk
numpy
oauth2client
openpyxl
outcome
paho-mqtt
pandas
Pillow
platformdirs
ply
prettytable
protobuf
pyasn1
pyasn1-modules
pycparser
pycryptodome
pydantic
PyJWT
pymongo==3.11.0
pyOpenSSL
pyparsing
pyrsistent
PySocks
python-dateutil
python-docx
python-dotenv
pytz
pytz-deprecation-shim
PyYAML
regex
repoze.lru
requests
requests-file
rfc3986
rsa
schedule
selenium
sgmllib3k
six
sniffio
sortedcontainers
soupsieve
starlette
statsd
tinysegmenter
tldextract
webdriver-manager
websockets
Werkzeug
ws4py
wsproto
xlrd
yarl
zipp
